feat: integrate Hibernate for database management and update Account and User models

+++ src/main/java/org/example/exception/AccountNotFoundException.java
@@ -1,7 +1,7 @@
 package org.example.exception;
 
 public class AccountNotFoundException extends BankBusinessException {
-    public AccountNotFoundException(int id) {
+    public AccountNotFoundException(long id) {
         super("Account with ID " + id + " not found");
     }
 }
