+++ src/main/java/org/example/exception/LackOfFundsException.java
@@ -1,7 +1,7 @@
 package org.example.exception;
 
 public class LackOfFundsException extends BankBusinessException {
-  public LackOfFundsException(int accountId) {
+  public LackOfFundsException(long accountId) {
     super("Error: lack of funds on account ID " + accountId );
   }
